+++ Mlib/Time/Fps/Measure_Fps.cpp
@@ -4,7 +4,7 @@
 
 using namespace Mlib;
 
-MeasureFps::MeasureFps(float alpha, unsigned int print_interval)
+MeasureFps::MeasureFps(double alpha, unsigned int print_interval)
     : last_dt_{ NAN }
     , mean_dt_{ 0 }
     , mad_dt_{ 0 }
@@ -18,7 +18,7 @@ MeasureFps::~MeasureFps() = default;
 void MeasureFps::tick() {
     auto current_time = std::chrono::steady_clock::now();
     if (last_time_ != std::chrono::steady_clock::time_point()) {
-        auto dt = std::chrono::duration<float>(current_time - last_time_).count();
+        auto dt = std::chrono::duration<double>(current_time - last_time_).count();
         mean_dt_ = (1 - alpha_) * mean_dt_ + alpha_ * dt;
         mad_dt_ = (1 - alpha_) * mad_dt_ + alpha_ * std::abs(mean_dt_ - dt);
         if (print_interval_ != UINT_MAX) {
@@ -40,26 +40,26 @@ void MeasureFps::reset() {
     *this = MeasureFps{ alpha_, print_counter_ };
 }
 
-float MeasureFps::mean_dt() const {
+double MeasureFps::mean_dt() const {
     return mean_dt_;
 }
 
-float MeasureFps::last_fps() const {
+double MeasureFps::last_fps() const {
     return 1 / last_dt_;
 }
 
-float MeasureFps::mean_fps() const {
+double MeasureFps::mean_fps() const {
     return 1 / mean_dt_;
 }
 
-float MeasureFps::mad_fps() const {
-    return (max_fps() - min_fps()) / 2.f;
+double MeasureFps::mad_fps() const {
+    return (max_fps() - min_fps()) / 2.;
 }
 
-float MeasureFps::min_fps() const {
+double MeasureFps::min_fps() const {
     return 1.f / (mean_dt_ + mad_dt_);
 }
 
-float MeasureFps::max_fps() const {
+double MeasureFps::max_fps() const {
     return 1.f / (mean_dt_ - mad_dt_);
 }
